fix header to be responsive again
